dependencies: pin duckdb CRAN pkg to 0.7.1-1
In an attempt to address
https://github.com/apache/arrow/issues/35917
https://github.com/voltrondata-labs/arrow-benchmarks-ci/issues/131

Assumption: the duckdb package move from
0.7.1-1 to 0.8.0 is responsible for result
differences. This patch will help to validate
or invalidate this assumption.

Let's see if this specific version pin breaks the
overall dependency resolution. If it works, we should
stay here and have a benchmark owner/author do the
investigation for what's required to perform a
duckdb bump.